Create README.md
1. When you buy Perreaoult Co NFT shares you will receive.

A. The NFT share representing the value of the original Physical artwork. 

B. One professional quality Art print of the Original Art work. 

C. One professional digital print quality copy of the original Art work.

D. The Owner of all shares of a represented original work will then be the owner of the Original as well as the collectable  NFT shares.